Move adapters into apollo module
This allows to register them in compilation time and avoid registering
them on runtime and baking them on apollo's codegen

More context:
https://kotlinlang.slack.com/archives/C01A6KM1SBZ/p1655892829229049?thread_ts=1655863366.491279&cid=C01A6KM1SBZ
