+++ apollo/src/main/kotlin/com/hedvig/android/typeadapter/PromiscuousLocalDateAdapter.kt
@@ -1,4 +1,4 @@
-package com.hedvig.app.util.apollo.adapter
+package com.hedvig.android.typeadapter
 
 import com.apollographql.apollo3.api.Adapter
 import com.apollographql.apollo3.api.CustomScalarAdapters
@@ -11,6 +11,7 @@ import java.time.LocalDate
  *
  * For example, the string "2011-01-02T05:19" will result in `LocalDate.of(2011, Month.JANUARY, 2)`
  */
+@Suppress("unused") // Used inside the `apollo {}` block inside build.gradle.kts
 object PromiscuousLocalDateAdapter : Adapter<LocalDate> {
     override fun fromJson(reader: JsonReader, customScalarAdapters: CustomScalarAdapters): LocalDate {
         val date = reader.nextString()!!
